Some improvements.
Use imperative in docstrings.

`LaTeX-toggle-Lua-editing-key' and `LaTeX-Lua-environments' now are
customizable variable.

Define `LaTeX-edit-Lua-code-parent-buffer' and
`LaTeX-edit-Lua-code-parent-buffer-point', so now `auctex-lua.el'
compiles silently.

In `LaTeX-mark-environment-contents', use
`LaTeX-find-matching-beginning' and `LaTeX-find-matching-end' to find
the beginning and the end of the current environment.  It’s a bit slower
than the current implementation, but it should be more robust.

Use `*%s [Lua]*' as temporary buffer name, see
http://wikemacs.org/index.php/Emacs_Terminology about the asterisk
naming convention.

Remap `save-buffer' to `LaTeX-edit-Lua-code-finish' instead of binding
the hard-coded `C-x C-s' to that function.

Remove the `LaTeX-edit-Lua-code-get-*' functions, just use the
corresponding variables.